build(Library all): add publish to github with  chagesets

+++ packages/fsa-database/src/models/excludedDirectories/deleteExcludedDirectoryName.ts
@@ -5,7 +5,8 @@ import { db } from '../../db/setup'
  * @param {number} id 
  * @returns {Promise<boolean>} A Promise of true or false when fulfilled
  */
-export default async function deleteExcludedDirectoryName(id: number): Promise<boolean> {
+export default async function deleteExcludedDirectoryName(id: number | undefined): Promise<boolean> {
+  if(!id) return false
   try {
   await db.excludedDirectories.delete(id)
   return true
